% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painbow_data.R
\docType{data}
\name{painbow_colors}
\alias{painbow_colors}
\title{An array of colors from XKCD's painbow comic}
\format{
An array with 192 hex color strings.
}
\source{
\url{https://xkcd.com/2537/}
}
\usage{
painbow_colors
}
\description{
This set of 192 colors is the Painbow. The colors were extracted using the pixels from the color bar on the right side of the comic.
}
\keyword{datasets}
