% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painbow_data.R
\docType{data}
\name{painbow_data}
\alias{painbow_data}
\title{A 2D heatmap from XKCD's painbow comic}
\format{
A data frame with 58,425 rows and 3 variables:
\describe{
  \item{x}{x-axis position}
  \item{y}{y-axis position}
  \item{value}{A value from 0 to 1}
}
}
\source{
\url{https://xkcd.com/2537/}
}
\usage{
painbow_data
}
\description{
A 2D heatmap from XKCD's painbow comic
}
\keyword{datasets}
